requirements.txt
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bidict==0.21.3
blis==0.7.7
catalogue==2.0.7
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
distlib==0.3.3
Django==3.2.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
easyocr==1.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
executing==0.8.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.3.0
Flask==0.10.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
fonttools==4.32.0
geographiclib==1.52
geopy==2.2.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
h11==0.13.0
hyperlink==21.0.0
idna==3.2
imageio==2.9.0
incremental==21.3.0
ipython==8.2.0
itemadapter==0.5.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.0.2
jmespath==1.0.0
joblib==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
locationtagger==0.0.1
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
mysql-connector-python==8.0.26
ndg-httpsclient==0.5.1
networkx==2.7.1
newspaper3k==0.2.8
nltk==3.7
numpy==1.21.3
opencv-python==4.5.1.48
opencv-python-headless==4.5.5.64
outcome==1.1.0
packaging==21.3
pandas==1.3.4
paramiko==2.7.2
parsel==1.6.0
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.0
preshed==3.0.6
prompt-toolkit==3.0.29
Protego==0.2.1
protobuf==3.20.0
pure-eval==0.2.2
pusher==1.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.20
pydantic==1.8.2
PyDispatcher==2.0.5
Pygments==2.11.2
pygorithm==1.0.4
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
Pysher @ git+https://github.com/nlsdfnbch/Pysher.git@8515f9db9a4dc995b012648c1aae4bf13334ba61
PySocks==1.7.1
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
PyWavelets==1.3.0
PyYAML==6.0
queuelib==1.6.2
regex==2022.4.24
requests==2.26.0
requests-file==1.5.1
scikit-image==0.19.2
scipy==1.8.0
Scrapy==2.6.1
selenium==4.1.3
service-identity==21.1.0
sgmllib3k==1.0.0
shadowsocks==2.8.2
simplejson==3.17.5
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.1
srsly==2.4.3
sshtunnel==0.4.0
stack-data==0.2.0
termcolor==1.1.0
thefuzz==0.19.0
thinc==8.0.15
tifffile==2022.3.25
tinysegmenter==0.3
tldextract==3.2.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typer==0.4.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.7
virtualenv==20.8.1
w3lib==1.22.0
wasabi==0.9.1
wcwidth==0.2.5
webdriver-manager==3.5.4
websocket-client==1.2.1
Werkzeug==2.0.2
ws2s-python==2.1.5
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
